Refactor FuturesStreamWriter to more generic LazyStreamWriter
